Modify test structure
